Move types.ts to type folder
