+++ electron/store.ts
@@ -1,6 +1,6 @@
 import ElectronStore from 'electron-store';
 
-import { RatingTemplateStore, RelicType } from '../src/types.ts';
+import { RatingTemplateStore, RelicType } from '@/type/types.ts';
 
 interface StoreData {
   relicMainStatsLevel: {
@@ -11,17 +11,6 @@ interface StoreData {
   };
   relicSubStatsScore: {
     [index: string]: { [index: string]: number | number[] };
-  };
-  relicRating: {
-    [index: string]: {
-      [index: string]: {
-        valuableSub: string[];
-        shouldLock: {
-          contain: string;
-          include: { [index: string]: string[] };
-        };
-      };
-    };
   };
   ratingTemplates: RatingTemplateStore;
 }
@@ -629,2782 +618,10 @@ const relicSubStatsScore = {
   },
 };
 
-const MusketeerCommonValuableSub = [RelicType.CRITRate, RelicType.CRITDMG, RelicType.SPD, RelicType.ATKPercentage];
-
-const PrisonerCommonValuableSub = [
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-  RelicType.ATKPercentage,
-  RelicType.SPD,
-  RelicType.EffectHitRate,
-];
-
-const GlamothCommonValuableSub = [
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-  RelicType.SPD,
-  RelicType.ATKPercentage,
-  RelicType.EffectHitRate,
-];
-
-const lzumoCommonValuableSub = [RelicType.CRITRate, RelicType.CRITDMG, RelicType.SPD, RelicType.ATKPercentage];
-
-const HertaCommonValuableSub = [
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-  RelicType.SPD,
-  RelicType.ATKPercentage,
-  RelicType.EffectHitRate,
-];
-
-const IPCCommonValuableSub = [
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-  RelicType.SPD,
-  RelicType.ATKPercentage,
-  RelicType.EffectHitRate,
-];
-
-const DiscipleCommonValuableSub = [RelicType.CRITRate, RelicType.CRITDMG, RelicType.SPD, RelicType.HPPercentage];
-
-const GrandDukeCommonValuableSub = [
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-  RelicType.SPD,
-  RelicType.ATKPercentage,
-  RelicType.ATKPercentage,
-  RelicType.DEFPercentage,
-  RelicType.HPPercentage,
-];
-
-const ElementalCommonValuableSub = [
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-  RelicType.SPD,
-  RelicType.ATKPercentage,
-  RelicType.ATKPercentage,
-  RelicType.DEFPercentage,
-  RelicType.HPPercentage,
-];
-
-const XianZhouCommonValuableSub = [RelicType.SPD, RelicType.ATKPercentage];
-
-const VonwacqCommonValuableSub = [RelicType.SPD, RelicType.ATKPercentage];
-
-const PenaconyCommonValuableSub = [RelicType.SPD, RelicType.ATKPercentage];
-
-const BelobogCommonValuableSub = [
-  RelicType.SPD,
-  RelicType.DEFPercentage,
-  RelicType.EffectHitRate,
-  RelicType.CRITRate,
-  RelicType.CRITDMG,
-];
-
-const TaliaCommonValuableSub = [RelicType.SPD, RelicType.BreakEffect];
-
-const SalsottoCommonValuableSub = [
-  RelicType.SPD,
-  RelicType.CRITDMG,
-  RelicType.CRITRate,
-  RelicType.ATKPercentage,
-  RelicType.DEFPercentage,
-  RelicType.HPPercentage,
-];
-
-const TaikiyanCommonValuableSub = [
-  RelicType.SPD,
-  RelicType.CRITDMG,
-  RelicType.CRITRate,
-  RelicType.ATKPercentage,
-  RelicType.DEFPercentage,
-  RelicType.HPPercentage,
-];
-
-const SigoniaCommonValuableSub = [
-  RelicType.SPD,
-  RelicType.CRITDMG,
-  RelicType.CRITRate,
-  RelicType.ATKPercentage,
-  RelicType.DEFPercentage,
-  RelicType.HPPercentage,
-];
-
-const ThiefCommonValuableSub = [RelicType.SPD, RelicType.BreakEffect, RelicType.ATKPercentage];
-
-const WatchmakerCommonValuableSub = [RelicType.SPD, RelicType.BreakEffect, RelicType.ATKPercentage];
-
-const PasserbyCommonValuableSub = [RelicType.SPD, RelicType.HPPercentage, RelicType.EffectRes, RelicType.ATKPercentage];
-
-const KnightCommonValuableSub = [
-  RelicType.SPD,
-  RelicType.DEFPercentage,
-  RelicType.EffectRes,
-  RelicType.CRITDMG,
-  RelicType.CRITRate,
-];
-
-const GuardCommonValuableSub = [
-  RelicType.SPD,
-  RelicType.DEFPercentage,
-  RelicType.EffectRes,
-  RelicType.EffectHitRate,
-  RelicType.HPPercentage,
-];
-
-const MessengerCommonValuableSub = [RelicType.SPD, RelicType.CRITRate, RelicType.CRITDMG, RelicType.ATKPercentage];
-
-const InsumousuCommonValuableSub = [RelicType.SPD, RelicType.EffectRes, RelicType.CRITDMG, RelicType.CRITRate];
-
-const relicRating = {
-  "Musketeer's Wild Wheat Felt Hat": {
-    [RelicType.HP]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Musketeer's Coarse Leather Gloves": {
-    [RelicType.ATK]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Musketeer's Wind-Hunting Shawl": {
-    [RelicType.CRITRate]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.OutgoingHealingBoost]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Musketeer's Rivets Riding Boots": {
-    [RelicType.SPD]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: MusketeerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Prisoner's Sealed Muzzle": {
-    [RelicType.HP]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Prisoner's Leadstone Shackles": {
-    [RelicType.ATK]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Prisoner's Repressive Straitjacket": {
-    [RelicType.CRITRate]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Prisoner's Restrictive Fetters": {
-    [RelicType.SPD]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: PrisonerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Glamoth's Iron Cavalry Regiment": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Glamoth's Silent Tombstone": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: GlamothCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "lzumo's Magatsu no Morokami": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "lzumo's Blades of Origin and End": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: lzumoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Herta's Space Station": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Herta's Wandering Trek": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: HertaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "The IPC's Mega HQ": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "The IPC's Trade Route": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: IPCCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Disciple's Prosthetic Eye": {
-    [RelicType.HP]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Disciple's Ingenium Hand": {
-    [RelicType.ATK]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Disciple's Dewy Feather Garb": {
-    [RelicType.CRITRate]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Disciple's Celestial Silk Sandals": {
-    [RelicType.SPD]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: DiscipleCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Grand Duke's Crown of Netherflame": {
-    [RelicType.HP]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Grand Duke's Gloves of Fieryfur": {
-    [RelicType.ATK]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Grand Duke's Robe of Grace": {
-    [RelicType.CRITRate]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Grand Duke's Ceremonial Boots": {
-    [RelicType.SPD]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: GrandDukeCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Pioneer's Heatproof Shell": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Pioneer's Lacuna Compass": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Pioneer's Sealed Lead Apron": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Pioneer's Starfaring Anchor": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Hunter's Artaius Hood": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Hunter's Lizard Gloves": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Hunter's Ice Dragon Cloak": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Hunter's Soft Elkskin Boots": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Champion's Headgear": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Champion's Heavy Gloves": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Champion's Chest Guard": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Champion's Fleetfoot Boots": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Firesmith's Obsidian Goggles": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Firesmith's Ring of Flame-Mastery": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Firesmith's Fireproof Apron": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Firesmith's Alloy Leg": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Genius's Ultraremote Sensing Visor": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Genius's Frequency Catcher": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Genius's Metafield Suit": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Genius's Gravity Walker": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Band's Polarized Sunglasses": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Band's Touring Bracelet": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Band's Leather Jacket With Studs": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Band's Ankle Boots With Rivets": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Eagle's Beaked Helmet": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Eagle's Soaring Ring": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Eagle's Winged Suit Harness": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Eagle's Quilted Puttees": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Wastelander's Breathing Mask": {
-    [RelicType.HP]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Wastelander's Desert Terminal": {
-    [RelicType.ATK]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Wastelander's Friar Robe": {
-    [RelicType.CRITRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Wastelander's Powered Greaves": {
-    [RelicType.SPD]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ElementalCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "The Xianzhou Luofu's Celestial Ark": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "The Xianzhou Luofu's Ambrosial Arbor Vines": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: XianZhouCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Vonwacq's Island of Birth": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Vonwacq's Islandic Coast": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: VonwacqCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Penacony's Grand Hotel": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Penacony's Dream-Seeking Tracks": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: PenaconyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Belobog's Fortress of Preservation": {
-    [RelicType.DEFPercentage]: {
-      valuableSub: BelobogCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Belobog's Iron Defense": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: BelobogCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: BelobogCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Talia's Nailscrap Town": {
-    [RelicType.DEFPercentage]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Talia's Exposed Electric Wire": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.BreakEffect]: {
-      valuableSub: TaliaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Salsotto's Moving City": {
-    [RelicType.DEFPercentage]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Salsotto's Terminator Line": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.BreakEffect]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: SalsottoCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  'Taikiyan Laser Stadium': {
-    [RelicType.DEFPercentage]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Taikiyan's Arclight Race Track": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.BreakEffect]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: TaikiyanCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Sigonia's Gaiathra Berth": {
-    [RelicType.DEFPercentage]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Sigonia's Knot of Cyclicality": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.BreakEffect]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: SigoniaCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Thief's Myriad-Faced Mask": {
-    [RelicType.HP]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Thief's Gloves With Prints": {
-    [RelicType.ATK]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Thief's Steel Grappling Hook": {
-    [RelicType.CRITRate]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.OutgoingHealingBoost]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Thief's Meteor Boots": {
-    [RelicType.SPD]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: ThiefCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Watchmaker's Telescoping Lens": {
-    [RelicType.HP]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Watchmaker's Fortuitous Wristwatch": {
-    [RelicType.ATK]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Watchmaker's Illusory Formal Suit": {
-    [RelicType.CRITRate]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.OutgoingHealingBoost]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Watchmaker's Dream-Concealing Dress Shoes": {
-    [RelicType.SPD]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: WatchmakerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Passerby's Rejuvenated Wooden Hairstick": {
-    [RelicType.HP]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Passerby's Roaming Dragon Bracer": {
-    [RelicType.ATK]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Passerby's Ragged Embroided Coat": {
-    [RelicType.OutgoingHealingBoost]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Passerby's Stygian Hiking Boots": {
-    [RelicType.SPD]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: PasserbyCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Knight's Forgiving Casque": {
-    [RelicType.HP]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Knight's Silent Oath Ring": {
-    [RelicType.ATK]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Knight's Solemn Breastplate": {
-    [RelicType.HPPercentage]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Knight's Iron Boots of Order": {
-    [RelicType.SPD]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: KnightCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Guard's Cast Iron Helmet": {
-    [RelicType.HP]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Guard's Shining Gauntlets": {
-    [RelicType.ATK]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Guard's Uniform of Old": {
-    [RelicType.HPPercentage]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Guard's Silver Greaves": {
-    [RelicType.SPD]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: GuardCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Messenger's Holovisor": {
-    [RelicType.HP]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Messenger's Transformative Arm": {
-    [RelicType.ATK]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Messenger's Secret Satchel": {
-    [RelicType.HPPercentage]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.EffectHitRate]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITRate]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.CRITDMG]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.OutgoingHealingBoost]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Messenger's Par-kool Sneakers": {
-    [RelicType.SPD]: {
-      valuableSub: MessengerCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Insumousu's Whalefall Ship": {
-    [RelicType.ATKPercentage]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.PhysicalDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.FireDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.IceDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.LightningDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.WindDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.QuantumDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ImaginaryDMGBoost]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Insumousu's Frayed Hawser": {
-    [RelicType.EnergyRegenerationRate]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.BreakEffect]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.DEFPercentage]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.ATKPercentage]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-    [RelicType.HPPercentage]: {
-      valuableSub: InsumousuCommonValuableSub,
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  'Valorous Mask of Northern Skies': {
-    [RelicType.HP]: {
-      valuableSub: [],
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  'Valorous Bracelet of Grappling Hooks': {
-    [RelicType.ATK]: {
-      valuableSub: [],
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  'Valorous Plate of Souring Flight': {},
-  'Valorous Greaves of Pursuing Hunt': {},
-  "Iron Cavalry's Homing Helm": {
-    [RelicType.HP]: {
-      valuableSub: [],
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Iron Cavalry's Crushing Wristguard": {
-    [RelicType.ATK]: {
-      valuableSub: [],
-      shouldLock: {
-        contain: '',
-        include: {},
-      },
-    },
-  },
-  "Iron Cavalry's Silvery Armor": {},
-  "Iron Cavalry's Skywalk Greaves": {},
-  "Forge's Lotus Lantern Wick": {},
-  "Forge's Heavenly Flamewheel Silk": {},
-  "Duran's Mechabeast Bridle": {},
-  "Duran's Tent of Golden Sky": {},
-};
-
 const store = new ElectronStore<StoreData>({
   defaults: {
     relicMainStatsLevel: relicMainStatsLevel,
     relicSubStatsScore: relicSubStatsScore,
-    relicRating: relicRating,
     ratingTemplates: {},
   },
 });
