+++ src/components/panel/relic-tool-panel/selector/relic-set-selector.tsx
@@ -11,7 +11,7 @@ import {
   CommandList,
 } from '@/components/ui/command.tsx';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover.tsx';
-import { RelicSets } from '@/types.ts';
+import { RelicSets } from '@/type/types.ts';
 
 type RelicSetSelectorProps = {
   selectedKeys: string[];
