+++ src/components/panel/scan-panel/sub-stats-drop-down.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { AllSubStats } from '../../../types.ts';
+import { AllSubStats } from '../../../type/types.ts';
 
 import { Checkbox } from '@/components/ui/checkbox.tsx';
 import {
