+++ src/hooks/use-relic-template-store.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 
-import { RatingRule, RatingTemplate, RatingTemplateStore } from '../types.ts';
+import { RatingRule, RatingTemplate, RatingTemplateStore } from '../type/types.ts';
 
 import {
   createOrUpdateRelicRatingRule,
