+++ src/type/types.ts
@@ -1,93 +1,93 @@
 import FuzzySet from 'fuzzyset.js';
 
-import char1 from '../src/assets/icon/character/img.png';
-import char2 from '../src/assets/icon/character/img_1.png';
-import char11 from '../src/assets/icon/character/img_10.png';
-import char12 from '../src/assets/icon/character/img_11.png';
-import char13 from '../src/assets/icon/character/img_12.png';
-import char14 from '../src/assets/icon/character/img_13.png';
-import char15 from '../src/assets/icon/character/img_14.png';
-import char16 from '../src/assets/icon/character/img_15.png';
-import char17 from '../src/assets/icon/character/img_16.png';
-import char18 from '../src/assets/icon/character/img_17.png';
-import char19 from '../src/assets/icon/character/img_18.png';
-import char20 from '../src/assets/icon/character/img_19.png';
-import char3 from '../src/assets/icon/character/img_2.png';
-import char21 from '../src/assets/icon/character/img_20.png';
-import char22 from '../src/assets/icon/character/img_21.png';
-import char23 from '../src/assets/icon/character/img_22.png';
-import char24 from '../src/assets/icon/character/img_23.png';
-import char25 from '../src/assets/icon/character/img_24.png';
-import char26 from '../src/assets/icon/character/img_25.png';
-import char27 from '../src/assets/icon/character/img_26.png';
-import char28 from '../src/assets/icon/character/img_27.png';
-import char29 from '../src/assets/icon/character/img_28.png';
-import char30 from '../src/assets/icon/character/img_29.png';
-import char4 from '../src/assets/icon/character/img_3.png';
-import char31 from '../src/assets/icon/character/img_30.png';
-import char32 from '../src/assets/icon/character/img_31.png';
-import char33 from '../src/assets/icon/character/img_32.png';
-import char34 from '../src/assets/icon/character/img_33.png';
-import char35 from '../src/assets/icon/character/img_34.png';
-import char36 from '../src/assets/icon/character/img_35.png';
-import char37 from '../src/assets/icon/character/img_36.png';
-import char38 from '../src/assets/icon/character/img_37.png';
-import char39 from '../src/assets/icon/character/img_38.png';
-import char40 from '../src/assets/icon/character/img_39.png';
-import char5 from '../src/assets/icon/character/img_4.png';
-import char41 from '../src/assets/icon/character/img_40.png';
-import char42 from '../src/assets/icon/character/img_41.png';
-import char43 from '../src/assets/icon/character/img_42.png';
-import char44 from '../src/assets/icon/character/img_43.png';
-import char45 from '../src/assets/icon/character/img_44.png';
-import char46 from '../src/assets/icon/character/img_45.png';
-import char47 from '../src/assets/icon/character/img_46.png';
-import char48 from '../src/assets/icon/character/img_47.png';
-import char49 from '../src/assets/icon/character/img_48.png';
-import char50 from '../src/assets/icon/character/img_49.png';
-import char6 from '../src/assets/icon/character/img_5.png';
-import char51 from '../src/assets/icon/character/img_50.png';
-import char52 from '../src/assets/icon/character/img_51.png';
-import char7 from '../src/assets/icon/character/img_6.png';
-import char8 from '../src/assets/icon/character/img_7.png';
-import char9 from '../src/assets/icon/character/img_8.png';
-import char10 from '../src/assets/icon/character/img_9.png';
-import image1 from '../src/assets/icon/relic/1.png';
-import image10 from '../src/assets/icon/relic/10.png';
-import image11 from '../src/assets/icon/relic/11.png';
-import image12 from '../src/assets/icon/relic/12.png';
-import image13 from '../src/assets/icon/relic/13.png';
-import image14 from '../src/assets/icon/relic/14.png';
-import image15 from '../src/assets/icon/relic/15.png';
-import image16 from '../src/assets/icon/relic/16.png';
-import image17 from '../src/assets/icon/relic/17.png';
-import image18 from '../src/assets/icon/relic/18.png';
-import image19 from '../src/assets/icon/relic/19.png';
-import image2 from '../src/assets/icon/relic/2.png';
-import image20 from '../src/assets/icon/relic/20.png';
-import image21 from '../src/assets/icon/relic/21.png';
-import image22 from '../src/assets/icon/relic/22.png';
-import image23 from '../src/assets/icon/relic/23.png';
-import image24 from '../src/assets/icon/relic/24.png';
-import image25 from '../src/assets/icon/relic/25.png';
-import image26 from '../src/assets/icon/relic/26.png';
-import image27 from '../src/assets/icon/relic/27.png';
-import image28 from '../src/assets/icon/relic/28.png';
-import image29 from '../src/assets/icon/relic/29.png';
-import image3 from '../src/assets/icon/relic/3.png';
-import image30 from '../src/assets/icon/relic/30.png';
-import image31 from '../src/assets/icon/relic/31.png';
-import image32 from '../src/assets/icon/relic/32.png';
-import image33 from '../src/assets/icon/relic/33.png';
-import image34 from '../src/assets/icon/relic/34.png';
-import image35 from '../src/assets/icon/relic/35.png';
-import image36 from '../src/assets/icon/relic/36.png';
-import image4 from '../src/assets/icon/relic/4.png';
-import image5 from '../src/assets/icon/relic/5.png';
-import image6 from '../src/assets/icon/relic/6.png';
-import image7 from '../src/assets/icon/relic/7.png';
-import image8 from '../src/assets/icon/relic/8.png';
-import image9 from '../src/assets/icon/relic/9.png';
+import char1 from '../assets/icon/character/img.png';
+import char2 from '../assets/icon/character/img_1.png';
+import char11 from '../assets/icon/character/img_10.png';
+import char12 from '../assets/icon/character/img_11.png';
+import char13 from '../assets/icon/character/img_12.png';
+import char14 from '../assets/icon/character/img_13.png';
+import char15 from '../assets/icon/character/img_14.png';
+import char16 from '../assets/icon/character/img_15.png';
+import char17 from '../assets/icon/character/img_16.png';
+import char18 from '../assets/icon/character/img_17.png';
+import char19 from '../assets/icon/character/img_18.png';
+import char20 from '../assets/icon/character/img_19.png';
+import char3 from '../assets/icon/character/img_2.png';
+import char21 from '../assets/icon/character/img_20.png';
+import char22 from '../assets/icon/character/img_21.png';
+import char23 from '../assets/icon/character/img_22.png';
+import char24 from '../assets/icon/character/img_23.png';
+import char25 from '../assets/icon/character/img_24.png';
+import char26 from '../assets/icon/character/img_25.png';
+import char27 from '../assets/icon/character/img_26.png';
+import char28 from '../assets/icon/character/img_27.png';
+import char29 from '../assets/icon/character/img_28.png';
+import char30 from '../assets/icon/character/img_29.png';
+import char4 from '../assets/icon/character/img_3.png';
+import char31 from '../assets/icon/character/img_30.png';
+import char32 from '../assets/icon/character/img_31.png';
+import char33 from '../assets/icon/character/img_32.png';
+import char34 from '../assets/icon/character/img_33.png';
+import char35 from '../assets/icon/character/img_34.png';
+import char36 from '../assets/icon/character/img_35.png';
+import char37 from '../assets/icon/character/img_36.png';
+import char38 from '../assets/icon/character/img_37.png';
+import char39 from '../assets/icon/character/img_38.png';
+import char40 from '../assets/icon/character/img_39.png';
+import char5 from '../assets/icon/character/img_4.png';
+import char41 from '../assets/icon/character/img_40.png';
+import char42 from '../assets/icon/character/img_41.png';
+import char43 from '../assets/icon/character/img_42.png';
+import char44 from '../assets/icon/character/img_43.png';
+import char45 from '../assets/icon/character/img_44.png';
+import char46 from '../assets/icon/character/img_45.png';
+import char47 from '../assets/icon/character/img_46.png';
+import char48 from '../assets/icon/character/img_47.png';
+import char49 from '../assets/icon/character/img_48.png';
+import char50 from '../assets/icon/character/img_49.png';
+import char6 from '../assets/icon/character/img_5.png';
+import char51 from '../assets/icon/character/img_50.png';
+import char52 from '../assets/icon/character/img_51.png';
+import char7 from '../assets/icon/character/img_6.png';
+import char8 from '../assets/icon/character/img_7.png';
+import char9 from '../assets/icon/character/img_8.png';
+import char10 from '../assets/icon/character/img_9.png';
+import image1 from '../assets/icon/relic/1.png';
+import image10 from '../assets/icon/relic/10.png';
+import image11 from '../assets/icon/relic/11.png';
+import image12 from '../assets/icon/relic/12.png';
+import image13 from '../assets/icon/relic/13.png';
+import image14 from '../assets/icon/relic/14.png';
+import image15 from '../assets/icon/relic/15.png';
+import image16 from '../assets/icon/relic/16.png';
+import image17 from '../assets/icon/relic/17.png';
+import image18 from '../assets/icon/relic/18.png';
+import image19 from '../assets/icon/relic/19.png';
+import image2 from '../assets/icon/relic/2.png';
+import image20 from '../assets/icon/relic/20.png';
+import image21 from '../assets/icon/relic/21.png';
+import image22 from '../assets/icon/relic/22.png';
+import image23 from '../assets/icon/relic/23.png';
+import image24 from '../assets/icon/relic/24.png';
+import image25 from '../assets/icon/relic/25.png';
+import image26 from '../assets/icon/relic/26.png';
+import image27 from '../assets/icon/relic/27.png';
+import image28 from '../assets/icon/relic/28.png';
+import image29 from '../assets/icon/relic/29.png';
+import image3 from '../assets/icon/relic/3.png';
+import image30 from '../assets/icon/relic/30.png';
+import image31 from '../assets/icon/relic/31.png';
+import image32 from '../assets/icon/relic/32.png';
+import image33 from '../assets/icon/relic/33.png';
+import image34 from '../assets/icon/relic/34.png';
+import image35 from '../assets/icon/relic/35.png';
+import image36 from '../assets/icon/relic/36.png';
+import image4 from '../assets/icon/relic/4.png';
+import image5 from '../assets/icon/relic/5.png';
+import image6 from '../assets/icon/relic/6.png';
+import image7 from '../assets/icon/relic/7.png';
+import image8 from '../assets/icon/relic/8.png';
+import image9 from '../assets/icon/relic/9.png';
 
 export type RatingRule = {
   setNames: string[];
@@ -145,16 +145,6 @@ export type OCRMainStatsResult = {
 export type OCRSubStatsResult = {
   result: RelicSubStats[] | null;
   error: string | null;
-};
-
-export type RelicRatingInfo = {
-  valuableSub: string[];
-  shouldLock: {
-    contain: string;
-    include: {
-      [key: string]: string[];
-    };
-  };
 };
 
 export enum RelicHeadMainStatsType {
